+++ changelog.rst
@@ -13,5 +13,7 @@ Changelog
 1.0.1
 =====
 
+- SQLite: Respect custom field parsers and formatters.
+- SQLite: Type conversion: Fix issue with incorrect conversion of field named ``name``.
 - Rowutils: Respect ``ro`` pragma in ``toRow`` proc.
 - Objutils: Respect ``ro`` pragma in ``fieldNames`` proc.